Refactor check_marathon_services_replication
This should be a noop in terms of functionality

Mostly moving things from check_marathon_services_replication to
monitoring_tools so that I will be able to re-use some of these
methods to create check_kubernetes_services_replication.

I also changed get_proxy_port_for_instance to take an InstanceConfig so
that I can reuse for k8s too. And removed the
read_all_registrations_for_service_instance and
read_registration_for_service_instance which I feel are not really
necessary to be separate functions.
