+++ paasta_tools/cli/cmds/get_latest_deployment.py
@@ -16,11 +16,11 @@ import sys
 
 from paasta_tools.cli.utils import lazy_choices_completer
 from paasta_tools.cli.utils import list_deploy_groups
-from paasta_tools.cli.utils import list_services
 from paasta_tools.cli.utils import PaastaColors
 from paasta_tools.cli.utils import validate_service_name
 from paasta_tools.deployment_utils import get_currently_deployed_sha
 from paasta_tools.utils import DEFAULT_SOA_DIR
+from paasta_tools.utils import list_services
 from paasta_tools.utils import paasta_print
