+++ paasta_tools/cli/cmds/remote_run.py
@@ -19,9 +19,9 @@ from shlex import quote
 from paasta_tools.cli.utils import lazy_choices_completer
 from paasta_tools.cli.utils import list_clusters
 from paasta_tools.cli.utils import list_instances
-from paasta_tools.cli.utils import list_services
 from paasta_tools.cli.utils import run_on_master
 from paasta_tools.utils import DEFAULT_SOA_DIR
+from paasta_tools.utils import list_services
 from paasta_tools.utils import load_system_paasta_config
 from paasta_tools.utils import paasta_print
 from paasta_tools.utils import PaastaColors
