+++ paasta_tools/cli/cmds/secret.py
@@ -16,9 +16,9 @@ import os
 import sys
 
 from paasta_tools.cli.utils import lazy_choices_completer
-from paasta_tools.cli.utils import list_services
 from paasta_tools.secret_tools import get_secret_provider
 from paasta_tools.utils import list_clusters
+from paasta_tools.utils import list_services
 from paasta_tools.utils import load_system_paasta_config
 from paasta_tools.utils import paasta_print
