+++ paasta_tools/monitoring_tools.py
@@ -27,7 +27,9 @@ import os
 import pysensu_yelp
 import service_configuration_lib
 
+from paasta_tools.utils import _log
 from paasta_tools.utils import DEFAULT_SOA_DIR
+from paasta_tools.utils import is_under_replicated
 from paasta_tools.utils import load_system_paasta_config
 from paasta_tools.utils import PaastaNotConfiguredError
 
@@ -235,3 +237,140 @@ def list_teams(**kwargs):
     team_data = _load_sensu_team_data()
     teams = set(team_data.get('team_data', {}).keys())
     return teams
+
+
+def send_replication_event(instance_config, status, output):
+    """Send an event to sensu via pysensu_yelp with the given information.
+
+    :param instance_config: an instance of LongRunningServiceConfig
+    :param status: The status to emit for this event
+    :param output: The output to emit for this event"""
+    # This function assumes the input is a string like "mumble.main"
+    monitoring_overrides = instance_config.get_monitoring()
+    if 'alert_after' not in monitoring_overrides:
+        monitoring_overrides['alert_after'] = '2m'
+    monitoring_overrides['check_every'] = '1m'
+    monitoring_overrides['runbook'] = get_runbook(
+        monitoring_overrides,
+        instance_config.service, soa_dir=instance_config.soa_dir,
+    )
+
+    check_name = (
+        'check_paasta_services_replication.%s' %
+        instance_config.job_id
+    )
+    send_event(
+        service=instance_config.service,
+        check_name=check_name,
+        overrides=monitoring_overrides,
+        status=status,
+        output=output,
+        soa_dir=instance_config.soa_dir,
+        cluster=instance_config.cluster,
+    )
+    _log(
+        service=instance_config.service,
+        line='Replication: %s' % output,
+        component='monitoring',
+        level='debug',
+        cluster=instance_config.cluster,
+        instance=instance_config.instance,
+    )
+
+
+def check_smartstack_replication_for_instance(
+    instance_config,
+    expected_count,
+    smartstack_replication_checker,
+):
+    """Check a set of namespaces to see if their number of available backends is too low,
+    emitting events to Sensu based on the fraction available and the thresholds defined in
+    the corresponding yelpsoa config.
+
+    :param instance_config: an instance of MarathonServiceConfig
+    :param smartstack_replication_checker: an instance of SmartstackReplicationChecker
+    """
+
+    crit_threshold = instance_config.get_replication_crit_percentage()
+
+    log.info('Checking instance %s in smartstack', instance_config.job_id)
+    smartstack_replication_info = \
+        smartstack_replication_checker.get_replication_for_instance(instance_config)
+
+    log.debug('Got smartstack replication info for %s: %s' %
+              (instance_config.job_id, smartstack_replication_info))
+
+    if len(smartstack_replication_info) == 0:
+        status = pysensu_yelp.Status.CRITICAL
+        output = (
+            'Service %s has no Smartstack replication info. Make sure the discover key in your smartstack.yaml '
+            'is valid!\n'
+        ) % instance_config.job_id
+        log.error(output)
+    else:
+        expected_count_per_location = int(expected_count / len(smartstack_replication_info))
+        output = ''
+        output_critical = ''
+        output_ok = ''
+        under_replication_per_location = []
+
+        for location, available_backends in sorted(smartstack_replication_info.items()):
+            num_available_in_location = available_backends.get(instance_config.job_id, 0)
+            under_replicated, ratio = is_under_replicated(
+                num_available_in_location, expected_count_per_location, crit_threshold,
+            )
+            if under_replicated:
+                output_critical += '- Service %s has %d out of %d expected instances in %s (CRITICAL: %d%%)\n' % (
+                    instance_config.job_id, num_available_in_location, expected_count_per_location, location, ratio,
+                )
+            else:
+                output_ok += '- Service %s has %d out of %d expected instances in %s (OK: %d%%)\n' % (
+                    instance_config.job_id, num_available_in_location, expected_count_per_location, location, ratio,
+                )
+            under_replication_per_location.append(under_replicated)
+
+        output += output_critical
+        if output_critical and output_ok:
+            output += '\n\n'
+            output += 'The following locations are OK:\n'
+        output += output_ok
+
+        if any(under_replication_per_location):
+            status = pysensu_yelp.Status.CRITICAL
+            output += (
+                "\n\n"
+                "What this alert means:\n"
+                "\n"
+                "  This replication alert means that a SmartStack powered loadbalancer (haproxy)\n"
+                "  doesn't have enough healthy backends. Not having enough healthy backends\n"
+                "  means that clients of that service will get 503s (http) or connection refused\n"
+                "  (tcp) when trying to connect to it.\n"
+                "\n"
+                "Reasons this might be happening:\n"
+                "\n"
+                "  The service may simply not have enough copies or it could simply be\n"
+                "  unhealthy in that location. There also may not be enough resources\n"
+                "  in the cluster to support the requested instance count.\n"
+                "\n"
+                "Things you can do:\n"
+                "\n"
+                "  * You can view the logs for the job with:\n"
+                "      paasta logs -s %(service)s -i %(instance)s -c %(cluster)s\n"
+                "\n"
+                "  * Fix the cause of the unhealthy service. Try running:\n"
+                "\n"
+                "      paasta status -s %(service)s -i %(instance)s -c %(cluster)s -vv\n"
+                "\n"
+                "  * Widen SmartStack discovery settings\n"
+                "  * Increase the instance count\n"
+                "\n"
+            ) % {
+                'service': instance_config.service,
+                'instance': instance_config.instance,
+                'cluster': instance_config.cluster,
+            }
+            log.error(output)
+        else:
+            status = pysensu_yelp.Status.OK
+            log.info(output)
+    send_replication_event(instance_config=instance_config, status=status, output=output)
