+++ tests/test_check_marathon_services_replication.py
@@ -15,7 +15,6 @@ from datetime import datetime
 from datetime import timedelta
 
 import mock
-import pysensu_yelp
 import pytest
 
 from paasta_tools import check_marathon_services_replication
@@ -41,387 +40,15 @@ def instance_config():
     return mock_instance_config
 
 
-def test_send_event_users_monitoring_tools_send_event_properly(instance_config):
-    fake_status = '999999'
-    fake_output = 'YOU DID IT'
-    instance_config.get_monitoring.return_value = {'fake_key': 'fake_value'}
-
-    expected_check_name = 'check_marathon_services_replication.%s' % instance_config.job_id
-    with mock.patch(
-        'paasta_tools.monitoring_tools.send_event', autospec=True,
-    ) as send_event_patch, mock.patch(
-        'paasta_tools.check_marathon_services_replication._log', autospec=True,
-    ), mock.patch(
-        'paasta_tools.check_marathon_services_replication.monitoring_tools.get_runbook',
-        autospec=True,
-        return_value='y/runbook',
-    ):
-        check_marathon_services_replication.send_event(
-            instance_config=instance_config,
-            status=fake_status,
-            output=fake_output,
-        )
-        send_event_patch.assert_called_once_with(
-            service=instance_config.service,
-            check_name=expected_check_name,
-            overrides={
-                'fake_key': 'fake_value', 'runbook': 'y/runbook',
-                'alert_after': '2m', 'check_every': '1m',
-            },
-            status=fake_status,
-            output=fake_output,
-            soa_dir=instance_config.soa_dir,
-            cluster=instance_config.cluster,
-        )
-
-
-def test_send_event_users_monitoring_tools_send_event_respects_alert_after(instance_config):
-    fake_status = '999999'
-    fake_output = 'YOU DID IT'
-    instance_config.get_monitoring.return_value = {'alert_after': '666m'}
-    expected_check_name = (
-        'check_marathon_services_replication.%s' %
-        instance_config.job_id
-    )
-    with mock.patch(
-        "paasta_tools.monitoring_tools.send_event", autospec=True,
-    ) as send_event_patch, mock.patch(
-        "paasta_tools.check_marathon_services_replication._log", autospec=True,
-    ), mock.patch(
-        'paasta_tools.check_marathon_services_replication.monitoring_tools.get_runbook',
-        autospec=True,
-        return_value='y/runbook',
-    ):
-        check_marathon_services_replication.send_event(
-            instance_config=instance_config,
-            status=fake_status,
-            output=fake_output,
-        )
-        send_event_patch.call_count == 1
-        send_event_patch.assert_called_once_with(
-            service=instance_config.service,
-            check_name=expected_check_name,
-            overrides={
-                'runbook': 'y/runbook',
-                'alert_after': '666m', 'check_every': '1m',
-            },
-            status=fake_status,
-            output=fake_output,
-            soa_dir=instance_config.soa_dir,
-            cluster=instance_config.cluster,
-        )
-
-
-def test_check_smartstack_replication_for_instance_ok_when_expecting_zero(instance_config):
-    expected_replication_count = 0
-    mock_smartstack_replication_checker = mock.Mock()
-    mock_smartstack_replication_checker.get_replication_for_instance.return_value = \
-        {'fake_region': {'test.main': 1, 'test.three': 4, 'test.four': 8}}
-
-    with mock.patch(
-        'paasta_tools.check_marathon_services_replication.send_event', autospec=True,
-    ) as mock_send_event:
-        check_marathon_services_replication.check_smartstack_replication_for_instance(
-            instance_config=instance_config,
-            expected_count=expected_replication_count,
-            smartstack_replication_checker=mock_smartstack_replication_checker,
-        )
-        mock_send_event.assert_called_once_with(
-            instance_config=instance_config,
-            status=pysensu_yelp.Status.OK,
-            output=mock.ANY,
-        )
-
-
-def test_check_smartstack_replication_for_instance_crit_when_absent(instance_config):
-    expected_replication_count = 8
-    mock_smartstack_replication_checker = mock.Mock()
-    mock_smartstack_replication_checker.get_replication_for_instance.return_value = \
-        {'fake_region': {'test.two': 1, 'test.three': 4, 'test.four': 8}}
-    with mock.patch(
-        'paasta_tools.check_marathon_services_replication.send_event', autospec=True,
-    ) as mock_send_event:
-        check_marathon_services_replication.check_smartstack_replication_for_instance(
-            instance_config=instance_config,
-            expected_count=expected_replication_count,
-            smartstack_replication_checker=mock_smartstack_replication_checker,
-        )
-        mock_send_event.assert_called_once_with(
-            instance_config=instance_config,
-            status=pysensu_yelp.Status.CRITICAL,
-            output=mock.ANY,
-        )
-
-
-def test_check_smartstack_replication_for_instance_crit_when_zero_replication(instance_config):
-    expected_replication_count = 8
-    mock_smartstack_replication_checker = mock.Mock()
-    mock_smartstack_replication_checker.get_replication_for_instance.return_value = \
-        {
-            'fake_region': {
-                'fake_service.fake_instance': 0,
-                'test.main': 8,
-                'test.fully_replicated': 8,
-            },
-        }
-    with mock.patch(
-        'paasta_tools.check_marathon_services_replication.send_event', autospec=True,
-    ) as mock_send_event:
-        check_marathon_services_replication.check_smartstack_replication_for_instance(
-            instance_config=instance_config,
-            expected_count=expected_replication_count,
-            smartstack_replication_checker=mock_smartstack_replication_checker,
-        )
-        mock_send_event.assert_called_once_with(
-            instance_config=instance_config,
-            status=pysensu_yelp.Status.CRITICAL,
-            output=mock.ANY,
-        )
-        _, send_event_kwargs = mock_send_event.call_args
-        alert_output = send_event_kwargs["output"]
-        assert ('Service {} has 0 out of 8 expected instances in fake_region'
-                .format(instance_config.job_id)) in alert_output
-        assert ('paasta status -s {} -i {} -c {} -vv'
-                .format(
-                    instance_config.service,
-                    instance_config.instance,
-                    instance_config.cluster,
-                )) in alert_output
-
-
-def test_check_smartstack_replication_for_instance_crit_when_low_replication(instance_config):
-    expected_replication_count = 8
-    mock_smartstack_replication_checker = mock.Mock()
-    mock_smartstack_replication_checker.get_replication_for_instance.return_value = \
-        {
-            'fake_region': {
-                'test.canary': 1,
-                'fake_service.fake_instance': 4,
-                'test.fully_replicated': 8,
-            },
-        }
-    with mock.patch(
-        'paasta_tools.check_marathon_services_replication.send_event', autospec=True,
-    ) as mock_send_event:
-        check_marathon_services_replication.check_smartstack_replication_for_instance(
-            instance_config=instance_config,
-            expected_count=expected_replication_count,
-            smartstack_replication_checker=mock_smartstack_replication_checker,
-        )
-        mock_send_event.assert_called_once_with(
-            instance_config=instance_config,
-            status=pysensu_yelp.Status.CRITICAL,
-            output=mock.ANY,
-        )
-        _, send_event_kwargs = mock_send_event.call_args
-        alert_output = send_event_kwargs["output"]
-        assert ('Service {} has 4 out of 8 expected instances in fake_region'
-                .format(instance_config.job_id)) in alert_output
-        assert ('paasta status -s {} -i {} -c {} -vv'
-                .format(
-                    instance_config.service,
-                    instance_config.instance,
-                    instance_config.cluster,
-                )) in alert_output
-
-
-def test_check_smartstack_replication_for_instance_ok_with_enough_replication(instance_config):
-    expected_replication_count = 8
-    mock_smartstack_replication_checker = mock.Mock()
-    mock_smartstack_replication_checker.get_replication_for_instance.return_value = \
-        {
-            'fake_region': {
-                'test.canary': 1,
-                'test.low_replication': 4,
-                'fake_service.fake_instance': 8,
-            },
-        }
-    with mock.patch(
-        'paasta_tools.check_marathon_services_replication.send_event', autospec=True,
-    ) as mock_send_event:
-        check_marathon_services_replication.check_smartstack_replication_for_instance(
-            instance_config=instance_config,
-            expected_count=expected_replication_count,
-            smartstack_replication_checker=mock_smartstack_replication_checker,
-        )
-        mock_send_event.assert_called_once_with(
-            instance_config=instance_config,
-            status=pysensu_yelp.Status.OK,
-            output=mock.ANY,
-        )
-        _, send_event_kwargs = mock_send_event.call_args
-        alert_output = send_event_kwargs["output"]
-        assert ('{} has 8 out of 8 expected instances in fake_region (OK: 100%)'
-                .format(instance_config.job_id)) in alert_output
-
-
-def test_check_smartstack_replication_for_instance_ok_with_enough_replication_multilocation(
-    instance_config,
-):
-    expected_replication_count = 2
-    mock_smartstack_replication_checker = mock.Mock()
-    mock_smartstack_replication_checker.get_replication_for_instance.return_value = \
-        {
-            'fake_region': {'fake_service.fake_instance': 1},
-            'fake_other_region': {'fake_service.fake_instance': 1},
-        }
-    with mock.patch(
-        'paasta_tools.check_marathon_services_replication.send_event', autospec=True,
-    ) as mock_send_event:
-        check_marathon_services_replication.check_smartstack_replication_for_instance(
-            instance_config=instance_config,
-            expected_count=expected_replication_count,
-            smartstack_replication_checker=mock_smartstack_replication_checker,
-        )
-        mock_send_event.assert_called_once_with(
-            instance_config=instance_config,
-            status=pysensu_yelp.Status.OK,
-            output=mock.ANY,
-        )
-        _, send_event_kwargs = mock_send_event.call_args
-        alert_output = send_event_kwargs["output"]
-        assert ("{} has 1 out of 1 expected instances in fake_region"
-                .format(instance_config.job_id)) in alert_output
-        assert ("{} has 1 out of 1 expected instances in fake_other_region"
-                .format(instance_config.job_id)) in alert_output
-
-
-def test_check_smartstack_replication_for_instance_crit_when_low_replication_multilocation(
-    instance_config,
-):
-    expected_replication_count = 2
-    mock_smartstack_replication_checker = mock.Mock()
-    mock_smartstack_replication_checker.get_replication_for_instance.return_value = \
-        {
-            'fake_region': {'fake_service.fake_instance': 1},
-            'fake_other_region': {'fake_service.fake_instance': 0},
-        }
-    with mock.patch(
-        'paasta_tools.check_marathon_services_replication.send_event', autospec=True,
-    ) as mock_send_event:
-        check_marathon_services_replication.check_smartstack_replication_for_instance(
-            instance_config=instance_config,
-            expected_count=expected_replication_count,
-            smartstack_replication_checker=mock_smartstack_replication_checker,
-        )
-        mock_send_event.assert_called_once_with(
-            instance_config=instance_config,
-            status=pysensu_yelp.Status.CRITICAL,
-            output=mock.ANY,
-        )
-        _, send_event_kwargs = mock_send_event.call_args
-        alert_output = send_event_kwargs["output"]
-        assert ("{} has 1 out of 1 expected instances in fake_region"
-                .format(instance_config.job_id)) in alert_output
-        assert ("{} has 0 out of 1 expected instances in fake_other_region"
-                .format(instance_config.job_id)) in alert_output
-        assert ("paasta status -s {} -i {} -c {} -vv"
-                .format(
-                    instance_config.service,
-                    instance_config.instance,
-                    instance_config.cluster,
-                )) in alert_output
-
-
-def test_check_smartstack_replication_for_instance_crit_when_zero_replication_multilocation(
-    instance_config,
-):
-    expected_replication_count = 2
-    mock_smartstack_replication_checker = mock.Mock()
-    mock_smartstack_replication_checker.get_replication_for_instance.return_value = \
-        {
-            'fake_region': {'fake_service.fake_instance': 0},
-            'fake_other_region': {'fake_service.fake_instance': 0},
-        }
-    with mock.patch(
-        'paasta_tools.check_marathon_services_replication.send_event', autospec=True,
-    ) as mock_send_event:
-        check_marathon_services_replication.check_smartstack_replication_for_instance(
-            instance_config=instance_config,
-            expected_count=expected_replication_count,
-            smartstack_replication_checker=mock_smartstack_replication_checker,
-        )
-        mock_send_event.assert_called_once_with(
-            instance_config=instance_config,
-            status=pysensu_yelp.Status.CRITICAL,
-            output=mock.ANY,
-        )
-        _, send_event_kwargs = mock_send_event.call_args
-        alert_output = send_event_kwargs["output"]
-        assert ("{} has 0 out of 1 expected instances in fake_region"
-                .format(instance_config.job_id)) in alert_output
-        assert ("{} has 0 out of 1 expected instances in fake_other_region"
-                .format(instance_config.job_id)) in alert_output
-        assert ("paasta status -s {} -i {} -c {} -vv"
-                .format(
-                    instance_config.service,
-                    instance_config.instance,
-                    instance_config.cluster,
-                )) in alert_output
-
-
-def test_check_smartstack_replication_for_instance_crit_when_missing_replication_multilocation(
-    instance_config,
-):
-    expected_replication_count = 2
-    mock_smartstack_replication_checker = mock.Mock()
-    mock_smartstack_replication_checker.get_replication_for_instance.return_value = \
-        {'fake_region': {'test.main': 0}, 'fake_other_region': {'test.main': 0}}
-    with mock.patch(
-        'paasta_tools.check_marathon_services_replication.send_event', autospec=True,
-    ) as mock_send_event:
-        check_marathon_services_replication.check_smartstack_replication_for_instance(
-            instance_config=instance_config,
-            expected_count=expected_replication_count,
-            smartstack_replication_checker=mock_smartstack_replication_checker,
-        )
-        mock_send_event.assert_called_once_with(
-            instance_config=instance_config,
-            status=pysensu_yelp.Status.CRITICAL,
-            output=mock.ANY,
-        )
-        _, send_event_kwargs = mock_send_event.call_args
-        alert_output = send_event_kwargs["output"]
-        assert ("{} has 0 out of 1 expected instances in fake_region"
-                .format(instance_config.job_id)) in alert_output
-        assert ("{} has 0 out of 1 expected instances in fake_other_region"
-                .format(instance_config.job_id)) in alert_output
-
-
-def test_check_smartstack_replication_for_instance_crit_when_no_smartstack_info(
-    instance_config,
-):
-    expected_replication_count = 2
-    mock_smartstack_replication_checker = mock.Mock()
-    mock_smartstack_replication_checker.get_replication_for_instance.return_value = {}
-    with mock.patch(
-        'paasta_tools.check_marathon_services_replication.send_event', autospec=True,
-    ) as mock_send_event:
-        check_marathon_services_replication.check_smartstack_replication_for_instance(
-            instance_config=instance_config,
-            expected_count=expected_replication_count,
-            smartstack_replication_checker=mock_smartstack_replication_checker,
-        )
-        mock_send_event.assert_called_once_with(
-            instance_config=instance_config,
-            status=pysensu_yelp.Status.CRITICAL,
-            output=mock.ANY,
-        )
-        _, send_event_kwargs = mock_send_event.call_args
-        alert_output = send_event_kwargs["output"]
-        assert ("{} has no Smartstack replication info."
-                .format(instance_config.job_id)) in alert_output
-
-
 def test_check_service_replication_for_normal_smartstack(instance_config):
     instance_config.get_instances.return_value = 100
     all_tasks = []
     with mock.patch(
-        'paasta_tools.marathon_tools.get_proxy_port_for_instance',
+        'paasta_tools.check_marathon_services_replication.get_proxy_port_for_instance',
         autospec=True,
         return_value=666,
     ), mock.patch(
-        'paasta_tools.check_marathon_services_replication.check_smartstack_replication_for_instance',
+        'paasta_tools.monitoring_tools.check_smartstack_replication_for_instance',
         autospec=True,
     ) as mock_check_smartstack_replication_for_service:
         check_marathon_services_replication.check_service_replication(
@@ -440,11 +67,11 @@ def test_check_service_replication_for_smartstack_with_different_namespace(insta
     instance_config.get_instances.return_value = 100
     all_tasks = []
     with mock.patch(
-        'paasta_tools.marathon_tools.get_proxy_port_for_instance',
+        'paasta_tools.check_marathon_services_replication.get_proxy_port_for_instance',
         autospec=True,
         return_value=666,
     ), mock.patch(
-        'paasta_tools.check_marathon_services_replication.check_smartstack_replication_for_instance',
+        'paasta_tools.monitoring_tools.check_smartstack_replication_for_instance',
         autospec=True,
     ) as mock_check_smartstack_replication_for_service, mock.patch(
         'paasta_tools.check_marathon_services_replication.check_healthy_marathon_tasks_for_service_instance',
@@ -468,7 +95,7 @@ def test_check_service_replication_for_non_smartstack(instance_config):
     instance_config.get_instances.return_value = 100
 
     with mock.patch(
-        'paasta_tools.marathon_tools.get_proxy_port_for_instance',
+        'paasta_tools.check_marathon_services_replication.get_proxy_port_for_instance',
         autospec=True,
         return_value=None,
     ), mock.patch(
@@ -539,11 +166,14 @@ def test_get_healthy_marathon_instances_for_short_app_id_considers_none_start_ti
     assert actual == 0
 
 
-@mock.patch('paasta_tools.check_marathon_services_replication.send_event_if_under_replication', autospec=True)
+@mock.patch(
+    'paasta_tools.check_marathon_services_replication.send_replication_event_if_under_replication',
+    autospec=True,
+)
 @mock.patch('paasta_tools.check_marathon_services_replication.filter_healthy_marathon_instances_for_short_app_id', autospec=True)  # noqa
 def test_check_healthy_marathon_tasks_for_service_instance(
     mock_healthy_instances,
-    mock_send_event_if_under_replication,
+    mock_send_replication_event_if_under_replication,
     instance_config,
 ):
     mock_healthy_instances.return_value = 2
@@ -552,18 +182,18 @@ def test_check_healthy_marathon_tasks_for_service_instance(
         expected_count=10,
         all_tasks=mock.Mock(),
     )
-    mock_send_event_if_under_replication.assert_called_once_with(
+    mock_send_replication_event_if_under_replication.assert_called_once_with(
         instance_config=instance_config,
         expected_count=10,
         num_available=2,
     )
 
 
-def test_send_event_if_under_replication_handles_0_expected(instance_config):
+def test_send_replication_event_if_under_replication_handles_0_expected(instance_config):
     with mock.patch(
-        'paasta_tools.check_marathon_services_replication.send_event', autospec=True,
+        'paasta_tools.monitoring_tools.send_replication_event', autospec=True,
     ) as mock_send_event:
-        check_marathon_services_replication.send_event_if_under_replication(
+        check_marathon_services_replication.send_replication_event_if_under_replication(
             instance_config=instance_config,
             expected_count=0,
             num_available=0,
@@ -579,11 +209,11 @@ def test_send_event_if_under_replication_handles_0_expected(instance_config):
                 .format(instance_config.job_id)) in alert_output
 
 
-def test_send_event_if_under_replication_good(instance_config):
+def test_send_replication_event_if_under_replication_good(instance_config):
     with mock.patch(
-        'paasta_tools.check_marathon_services_replication.send_event', autospec=True,
+        'paasta_tools.monitoring_tools.send_replication_event', autospec=True,
     ) as mock_send_event:
-        check_marathon_services_replication.send_event_if_under_replication(
+        check_marathon_services_replication.send_replication_event_if_under_replication(
             instance_config=instance_config,
             expected_count=100,
             num_available=100,
@@ -599,11 +229,11 @@ def test_send_event_if_under_replication_good(instance_config):
                 .format(instance_config.job_id)) in alert_output
 
 
-def test_send_event_if_under_replication_critical(instance_config):
+def test_send_replication_event_if_under_replication_critical(instance_config):
     with mock.patch(
-        'paasta_tools.check_marathon_services_replication.send_event', autospec=True,
+        'paasta_tools.monitoring_tools.send_replication_event', autospec=True,
     ) as mock_send_event:
-        check_marathon_services_replication.send_event_if_under_replication(
+        check_marathon_services_replication.send_replication_event_if_under_replication(
             instance_config=instance_config,
             expected_count=100,
             num_available=89,
